Arreglar que el nivel se diera la vuelta a 0 a los ~10,6 dias (v1.8)
level() devolvia uint8_t y, con MINUTES_PER_LEVEL=60, el nivel son las horas:
a las 256 h se daba la vuelta a 0. No es un caso rebuscado, se alcanza jugando
normal, porque "quedaros juntos" permite posponer la despedida indefinidamente
y hay gente con Pokemon por encima del nivel 100.

Ahora es uint16_t y se topa en 999 en vez de desbordar. De paso, dos sitios que
truncaban por la misma via:

- calcStat() recibia el nivel en uint8_t, asi que las estadisticas habrian
  vuelto a desbordar aunque level() ya no lo hiciera.
- canEvolveNow() comparaba contra (uint8_t)(evolveLevel + careMistakes), que
  con muchos descuidos se daba la vuelta y adelantaba la evolucion. Ademas la
  ficha ya calculaba ese mismo umbral en int, asi que la pantalla y la
  comprobacion real podian discrepar. Ahora coinciden.

evoDeclinedLv pasa tambien a uint16_t (no se persiste en NVS, no hay migracion).

Diagnostico y arreglo de ajtudela, de su rama refactor/improvments.

+++ pet.cpp
@@ -321,7 +321,7 @@ void Pet::rename(const char *name) {
   save();
 }
 
-static uint16_t calcStat(uint8_t base, uint8_t gene, uint8_t lvl, uint8_t tr) {
+static uint16_t calcStat(uint8_t base, uint8_t gene, uint16_t lvl, uint8_t tr) {
   return (uint16_t)base * gene / 100 + lvl + tr;
 }
 
@@ -412,7 +412,7 @@ bool Pet::canEvolveNow() const {
   if (isEgg() || sleeping || ceremony != CER_NONE) return false;
   const DexEntry &d = DEX_TBL[speciesId];
   if (d.evolvesTo == 0) return false;
-  return level() >= (uint8_t)(d.evolveLevel + careMistakes) && lowestStat() >= 40;
+  return level() >= (uint16_t)d.evolveLevel + careMistakes && lowestStat() >= 40;
 }
 
 void Pet::evolve() {
